refactor: 중복되어 있던 category에 대한 api 파일 및 타입 변경
