+++ types/Category.ts
@@ -9,4 +9,9 @@ export interface CategoryDto {
     categoryId: number;
     categoryName:string;
     categoryImageUrl:string;
-} 
+}
+
+export interface CategoryName {
+    categoryId: number;
+    categoryName: string;
+}
