--- types/CategoryName.ts
@@ -1,4 +1,0 @@
-export interface CategoryName {
-    categoryId: number;
-    categoryName: string;
-}
